refactor(kirra-verifier-pg): split PgVerifierStore seams into per-store modules
lib.rs was 1534 lines: PgVerifierStore with 22 impl blocks (one trait impl per
storage seam, several with a small inherent row-mapping helper impl + a
*_COLUMNS const), all in the crate root. Move each seam into its own submodule
(mirroring the per-table layout kirra-persistence uses on the SQLite side):

  nodes.rs (NodeStore) · epoch.rs (EpochFence) · posture.rs
  (PostureEngineStateStore) · federation.rs (FederationStore) · operators.rs
  (OperatorStore) · principals.rs (PrincipalStore) · cert_principals.rs
  (CertPrincipalStore) · fabric.rs (FabricAssetStore) · ota_campaigns.rs
  (OtaCampaignStore) · av_subsystem.rs (AvSubsystemStore)

lib.rs keeps the schema/migration engine (BASELINE_DDL, PG_MIGRATIONS,
PG_SCHEMA_VERSION), LivePgExecutor, PgStoreError, and the PgVerifierStore struct
+ core inherent impl (connect/from_client/initialize/schema_version/lock/
row_to_node). lib.rs 1534 -> 475 lines.

Behaviour-preserving — pure relocation. The only visibility change is
`pub(crate)` on the two shared helpers `lock` and `row_to_node` so the split
impl blocks reach them (each store's own row-mapping helpers move with it and
stay private). No public API change; all trait impls remain on PgVerifierStore.

Verified in the workspace-detached crate (the postgres-conformance lane's
`cd crates/kirra-verifier-pg && cargo ...`): lib + tests/live_pg.rs compile,
clippy clean, fmt clean. De-monolith #3, after #978/#979.
