+++ crates/kirra-verifier-pg/src/lib.rs
@@ -440,11 +440,11 @@ impl PgVerifierStore {
         Ok(rows.first().map(|r| r.get::<_, i64>(0)).unwrap_or(0))
     }
 
-    fn lock(&self) -> std::sync::MutexGuard<'_, postgres::Client> {
+    pub(crate) fn lock(&self) -> std::sync::MutexGuard<'_, postgres::Client> {
         self.client.lock().unwrap_or_else(|e| e.into_inner())
     }
 
-    fn row_to_node(row: &postgres::Row) -> RegisteredNode {
+    pub(crate) fn row_to_node(row: &postgres::Row) -> RegisteredNode {
         let status_json: String = row.get(1);
         // Same corrupt-status fallback as the SQLite backend: an undecodable
         // status is Unknown (fail toward "not trusted"), never a panic.
@@ -463,1072 +463,13 @@ impl PgVerifierStore {
     }
 }
 
-const NODE_COLUMNS: &str = "node_id, status_json, registered_at_ms, last_trust_update_ms, \
-                            ak_public_pem, expected_pcr16_digest_hex, site, firmware_version";
-
-impl NodeStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn save_node(&self, node: &RegisteredNode) -> Result<(), PgStoreError> {
-        let status_json = serde_json::to_string(&node.status).map_err(PgStoreError::Encode)?;
-        self.lock().execute(
-            "INSERT INTO nodes (node_id, status_json, registered_at_ms, last_trust_update_ms, \
-                                ak_public_pem, expected_pcr16_digest_hex, site, firmware_version) \
-             VALUES ($1, $2, $3, $4, $5, $6, $7, $8) \
-             ON CONFLICT (node_id) DO UPDATE SET \
-                 status_json = EXCLUDED.status_json, \
-                 registered_at_ms = EXCLUDED.registered_at_ms, \
-                 last_trust_update_ms = EXCLUDED.last_trust_update_ms, \
-                 ak_public_pem = EXCLUDED.ak_public_pem, \
-                 expected_pcr16_digest_hex = EXCLUDED.expected_pcr16_digest_hex, \
-                 site = EXCLUDED.site, \
-                 firmware_version = EXCLUDED.firmware_version",
-            &[
-                &node.node_id,
-                &status_json,
-                &(node.registered_at_ms as i64),
-                &(node.last_trust_update_ms as i64),
-                &node.ak_public_pem,
-                &node.expected_pcr16_digest_hex,
-                &node.site,
-                &node.firmware_version,
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn load_node(&self, node_id: &str) -> Result<Option<RegisteredNode>, PgStoreError> {
-        let row = self.lock().query_opt(
-            &format!("SELECT {NODE_COLUMNS} FROM nodes WHERE node_id = $1"),
-            &[&node_id],
-        )?;
-        Ok(row.as_ref().map(Self::row_to_node))
-    }
-
-    fn load_nodes(&self) -> Result<Vec<RegisteredNode>, PgStoreError> {
-        let rows = self
-            .lock()
-            .query(&format!("SELECT {NODE_COLUMNS} FROM nodes"), &[])?;
-        Ok(rows.iter().map(Self::row_to_node).collect())
-    }
-
-    fn node_exists(&self, node_id: &str) -> Result<bool, PgStoreError> {
-        let row = self
-            .lock()
-            .query_one("SELECT COUNT(*) FROM nodes WHERE node_id = $1", &[&node_id])?;
-        Ok(row.get::<_, i64>(0) > 0)
-    }
-
-    fn count_nodes(&self) -> Result<i64, PgStoreError> {
-        let row = self.lock().query_one("SELECT COUNT(*) FROM nodes", &[])?;
-        Ok(row.get::<_, i64>(0))
-    }
-}
-
-impl EpochFence for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn current_epoch(&self) -> Result<u64, PgStoreError> {
-        // `query_one` errors when the singleton row is absent — the read path
-        // is fail-closed exactly like the SQLite backend's `query_row`.
-        let row = self
-            .lock()
-            .query_one("SELECT epoch FROM ha_state WHERE id = 1", &[])?;
-        Ok(row.get::<_, i64>(0) as u64)
-    }
-
-    fn current_active_holder(&self) -> Result<(u64, Option<String>), PgStoreError> {
-        let row = self.lock().query_one(
-            "SELECT epoch, active_instance_id FROM ha_state WHERE id = 1",
-            &[],
-        )?;
-        Ok((row.get::<_, i64>(0) as u64, row.get(1)))
-    }
-
-    fn try_claim_epoch(
-        &mut self,
-        observed: u64,
-        instance_id: &str,
-        now_ms: u64,
-    ) -> Result<Option<u64>, PgStoreError> {
-        // The SAME rows-affected compare-and-set as the SQLite backend: two
-        // racers reading the same `observed` serialize at the row lock and
-        // exactly one sees `rows == 1`.
-        let n = self.lock().execute(
-            "UPDATE ha_state SET epoch = epoch + 1, active_instance_id = $2, updated_at_ms = $3 \
-             WHERE id = 1 AND epoch = $1",
-            &[&(observed as i64), &instance_id, &(now_ms as i64)],
-        )?;
-        Ok(if n == 1 { Some(observed + 1) } else { None })
-    }
-
-    fn assert_actuator_epoch_held(&mut self, held_epoch: u64) -> Result<(), FenceError> {
-        // The cross-backend constraint recorded on the trait, realized the
-        // Postgres way: a transaction whose `SELECT … FOR UPDATE` takes the
-        // `ha_state` row lock. While this assertion transaction is open a
-        // competing `try_claim_epoch` serializes behind it; if a competing
-        // claim already landed, the locked read observes the newer epoch and
-        // the fence rejects before any actuator response is issued. Fail
-        // closed on EVERY failure: transaction/read errors → `EpochUnreadable`;
-        // `held == 0` or any mismatch → `EpochSuperseded`. The transaction
-        // rolls back on drop, so the reject path never commits anything.
-        let mut guard = self.lock();
-        let mut tx = guard
-            .transaction()
-            .map_err(|_| FenceError::EpochUnreadable)?;
-        let row = tx
-            .query_opt("SELECT epoch FROM ha_state WHERE id = 1 FOR UPDATE", &[])
-            .map_err(|_| FenceError::EpochUnreadable)?;
-        let durable = match row {
-            Some(r) => r.get::<_, i64>(0) as u64,
-            // Singleton row absent — never authorize blind.
-            None => return Err(FenceError::EpochUnreadable),
-        };
-        if held_epoch == 0 || durable != held_epoch {
-            return Err(FenceError::EpochSuperseded {
-                held: held_epoch,
-                durable,
-            });
-        }
-        tx.commit().map_err(|_| FenceError::EpochUnreadable)
-    }
-}
-
-impl PostureEngineStateStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn load_last_generation(&self) -> Result<u64, PgStoreError> {
-        let row = self.lock().query_opt(
-            "SELECT value FROM posture_engine_state WHERE key = 'last_generation'",
-            &[],
-        )?;
-        match row {
-            // Genuinely absent row → 0 (fresh store), like SQLite's
-            // `QueryReturnedNoRows => Ok(0)`.
-            None => Ok(0),
-            // Present but non-numeric OR out-of-domain (`>= i64::MAX`) → CORRUPTION,
-            // fail closed. Same verdict as `VerifierStore::load_last_generation`.
-            Some(r) => {
-                let s: String = r.get(0);
-                let parsed = s
-                    .parse::<u64>()
-                    .map_err(|_| PgStoreError::CorruptGeneration(s.clone()))?;
-                if parsed >= i64::MAX as u64 {
-                    return Err(PgStoreError::CorruptGeneration(s));
-                }
-                Ok(parsed)
-            }
-        }
-    }
-
-    fn save_last_generation(&self, generation: u64) -> Result<bool, PgStoreError> {
-        // Single-statement conditional upsert — ATOMIC, so it is race-safe across
-        // connections exactly like the SQLite backend's monotonic upsert: there is no
-        // read-then-write window in which two racers could both read the old value and
-        // a lower generation clobber a higher one. The `WHERE` gates only the DO UPDATE
-        // (conflict) path; a first insert is unconditional. `rows == 1` on insert or an
-        // accepted strict advance, `0` when the guard rejects a stale/equal generation.
-        //
-        // The `CASE` guards the cast of the STORED value: a non-numeric existing value
-        // counts as 0, so a positive generation OVERWRITES it — the exact
-        // heal-on-corrupt-save parity SQLite gives via `CAST('garbage') = 0` (and the
-        // in-memory backend via `parse().unwrap_or(0)`). Only `load_last_generation`
-        // fails closed on a corrupt high-water; `save` heals, matching every backend.
-        //
-        // The comparison is `::numeric` (arbitrary precision), NOT `::bigint`: a stored
-        // value that is all-digits but out of the i64/u64 domain (a huge digit string
-        // written via a blind `save_engine_state`) would OVERFLOW a `::bigint` cast and
-        // re-raise the very crash this heals. `::numeric` never overflows — such a value
-        // simply compares greater than any real generation, so the monotonic guard keeps
-        // it (never a crash). `EXCLUDED.value` is `$1` (a valid u64 string), likewise
-        // overflow-free as `numeric`.
-        let n = self.lock().execute(
-            "INSERT INTO posture_engine_state (key, value) VALUES ('last_generation', $1) \
-             ON CONFLICT (key) DO UPDATE SET value = EXCLUDED.value \
-             WHERE (CASE WHEN posture_engine_state.value ~ '^[0-9]+$' \
-                         THEN posture_engine_state.value::numeric ELSE 0 END) \
-                   < EXCLUDED.value::numeric",
-            &[&generation.to_string()],
-        )?;
-        Ok(n == 1)
-    }
-
-    fn load_engine_state(&self, key: &str) -> Result<Option<String>, PgStoreError> {
-        let row = self.lock().query_opt(
-            "SELECT value FROM posture_engine_state WHERE key = $1",
-            &[&key],
-        )?;
-        Ok(row.map(|r| r.get::<_, String>(0)))
-    }
-
-    fn save_engine_state(&self, key: &str, value: &str) -> Result<(), PgStoreError> {
-        // Blind upsert (INSERT-OR-REPLACE) — NOT the monotonic guard.
-        self.lock().execute(
-            "INSERT INTO posture_engine_state (key, value) VALUES ($1, $2) \
-             ON CONFLICT (key) DO UPDATE SET value = EXCLUDED.value",
-            &[&key, &value],
-        )?;
-        Ok(())
-    }
-}
-
-impl FederationStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn save_trusted_federation_controller(
-        &self,
-        controller_id: &str,
-        public_key_b64: &str,
-        registered_at_ms: u64,
-    ) -> Result<(), PgStoreError> {
-        // Fail-closed on an out-of-domain timestamp rather than wrapping a `u64` to a
-        // negative `BIGINT` (bounded in practice — overflows i64 only past year 292M).
-        let reg_ms = i64::try_from(registered_at_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "registered_at_ms",
-            value: registered_at_ms,
-        })?;
-        // Upsert by controller_id (SQLite: INSERT OR REPLACE) — re-registering a
-        // controller overwrites its key.
-        self.lock().execute(
-            "INSERT INTO trusted_federation_controllers \
-                 (controller_id, public_key_b64, registered_at_ms) \
-             VALUES ($1, $2, $3) \
-             ON CONFLICT (controller_id) DO UPDATE SET \
-                 public_key_b64 = EXCLUDED.public_key_b64, \
-                 registered_at_ms = EXCLUDED.registered_at_ms",
-            &[&controller_id, &public_key_b64, &reg_ms],
-        )?;
-        Ok(())
-    }
-
-    fn load_trusted_federation_controller_key(
-        &self,
-        controller_id: &str,
-    ) -> Result<Option<String>, PgStoreError> {
-        let row = self.lock().query_opt(
-            "SELECT public_key_b64 FROM trusted_federation_controllers WHERE controller_id = $1",
-            &[&controller_id],
-        )?;
-        Ok(row.map(|r| r.get::<_, String>(0)))
-    }
-
-    fn has_seen_federation_nonce(&self, nonce_hex: &str) -> Result<bool, PgStoreError> {
-        let row = self.lock().query_one(
-            "SELECT COUNT(*) FROM federation_report_nonces WHERE nonce_hex = $1",
-            &[&nonce_hex],
-        )?;
-        Ok(row.get::<_, i64>(0) > 0)
-    }
-
-    fn burn_federation_nonce(&self, nonce_hex: &str) -> Result<bool, PgStoreError> {
-        // Atomic single-use claim: `ON CONFLICT DO NOTHING` is the Postgres
-        // `INSERT OR IGNORE` — `rows == 1` iff the nonce was newly recorded (first
-        // use → proceed), `0` on a replay (already present → reject). No
-        // check-then-act window; the PK conflict decides. `seen_at_ms` is diagnostic
-        // only (correctness rests on the PK, never the clock), and the source label
-        // mirrors the SQLite burn path.
-        let seen_at_ms = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_millis() as i64)
-            .unwrap_or(0);
-        let n = self.lock().execute(
-            "INSERT INTO federation_report_nonces (nonce_hex, source_controller_id, seen_at_ms) \
-             VALUES ($1, 'fleet-grant-lane', $2) \
-             ON CONFLICT (nonce_hex) DO NOTHING",
-            &[&nonce_hex, &seen_at_ms],
-        )?;
-        Ok(n == 1)
-    }
-
-    fn industrial_seq_check_and_advance(
-        &self,
-        source_id: &str,
-        sequence: u64,
-        now_ms: u64,
-    ) -> Result<bool, PgStoreError> {
-        // Fail-closed on an out-of-domain sequence/timestamp: a `u64` wrapped to a
-        // negative `BIGINT` would DEFEAT the gate (a wrapped-negative high-water lets a
-        // later smaller sequence compare "greater"), so refuse rather than wrap.
-        let seq = i64::try_from(sequence).map_err(|_| PgStoreError::OutOfDomain {
-            field: "sequence",
-            value: sequence,
-        })?;
-        let now = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // Atomic per-source strictly-advancing gate — the SAME conditional compare-
-        // and-set as the SQLite backend: a first message from a new source inserts
-        // (baseline accept), an existing source advances ONLY on a strictly greater
-        // sequence (the `WHERE` gates the DO UPDATE), and a replay/regress no-ops.
-        // `rows == 1` on accept, `0` on reject. Race-safe at the row lock.
-        let n = self.lock().execute(
-            "INSERT INTO industrial_message_seq (source_id, last_sequence, last_seen_ms) \
-             VALUES ($1, $2, $3) \
-             ON CONFLICT (source_id) DO UPDATE SET \
-                 last_sequence = EXCLUDED.last_sequence, \
-                 last_seen_ms = EXCLUDED.last_seen_ms \
-             WHERE EXCLUDED.last_sequence > industrial_message_seq.last_sequence",
-            &[&source_id, &seq, &now],
-        )?;
-        Ok(n == 1)
-    }
-}
-
-impl PgVerifierStore {
-    fn row_to_operator(row: &postgres::Row) -> OperatorRecord {
-        OperatorRecord {
-            operator_id: row.get(0),
-            pubkey_pem: row.get(1),
-            registered_at_ms: row.get::<_, i64>(2) as u64,
-            revoked_at_ms: row.get::<_, Option<i64>>(3).map(|v| v as u64),
-        }
-    }
-}
-
-impl OperatorStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn register_operator(
-        &mut self,
-        operator_id: &str,
-        pubkey_pem: &str,
-        now_ms: u64,
-    ) -> Result<(), PgStoreError> {
-        // Fail-closed on an out-of-domain timestamp (see FederationStore).
-        let reg_ms = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // Register or rotate: overwrite the key and CLEAR any prior revocation (a
-        // fresh key reactivates), matching the SQLite upsert.
-        self.lock().execute(
-            "INSERT INTO operators (operator_id, pubkey_pem, registered_at_ms, revoked_at_ms) \
-             VALUES ($1, $2, $3, NULL) \
-             ON CONFLICT (operator_id) DO UPDATE SET \
-                 pubkey_pem = EXCLUDED.pubkey_pem, \
-                 registered_at_ms = EXCLUDED.registered_at_ms, \
-                 revoked_at_ms = NULL",
-            &[&operator_id, &pubkey_pem, &reg_ms],
-        )?;
-        Ok(())
-    }
-
-    fn revoke_operator(&mut self, operator_id: &str, now_ms: u64) -> Result<bool, PgStoreError> {
-        let rev_ms = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // Conditional update — `rows == 1` iff an ACTIVE operator transitioned to
-        // revoked; `0` if absent or already revoked.
-        let n = self.lock().execute(
-            "UPDATE operators SET revoked_at_ms = $2 \
-             WHERE operator_id = $1 AND revoked_at_ms IS NULL",
-            &[&operator_id, &rev_ms],
-        )?;
-        Ok(n > 0)
-    }
-
-    fn load_operator(&self, operator_id: &str) -> Result<Option<OperatorRecord>, PgStoreError> {
-        let row = self.lock().query_opt(
-            "SELECT operator_id, pubkey_pem, registered_at_ms, revoked_at_ms \
-             FROM operators WHERE operator_id = $1",
-            &[&operator_id],
-        )?;
-        Ok(row.as_ref().map(Self::row_to_operator))
-    }
-
-    fn load_operators(&self) -> Result<Vec<OperatorRecord>, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT operator_id, pubkey_pem, registered_at_ms, revoked_at_ms \
-             FROM operators ORDER BY operator_id",
-            &[],
-        )?;
-        Ok(rows.iter().map(Self::row_to_operator).collect())
-    }
-}
-
-impl PgVerifierStore {
-    fn row_to_api_principal(row: &postgres::Row) -> ApiPrincipalRecord {
-        ApiPrincipalRecord {
-            principal_id: row.get(0),
-            role: row.get(1),
-            created_at_ms: row.get::<_, i64>(2) as u64,
-            revoked_at_ms: row.get::<_, Option<i64>>(3).map(|v| v as u64),
-        }
-    }
-}
-
-impl PrincipalStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn register_api_principal(
-        &mut self,
-        principal_id: &str,
-        token_sha256: &str,
-        role: &str,
-        now_ms: u64,
-    ) -> Result<(), PgStoreError> {
-        let created_ms = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // Register / rotate: overwrite token hash + role and CLEAR revocation. A
-        // `token_sha256` already held by a DIFFERENT principal violates the UNIQUE
-        // constraint (only the principal_id conflict is handled by ON CONFLICT), so
-        // it surfaces as a driver error — the fail-closed "one token, one principal"
-        // guarantee, exactly like the SQLite backend.
-        self.lock().execute(
-            "INSERT INTO api_principals \
-                 (principal_id, token_sha256, role, created_at_ms, revoked_at_ms) \
-             VALUES ($1, $2, $3, $4, NULL) \
-             ON CONFLICT (principal_id) DO UPDATE SET \
-                 token_sha256  = EXCLUDED.token_sha256, \
-                 role          = EXCLUDED.role, \
-                 created_at_ms = EXCLUDED.created_at_ms, \
-                 revoked_at_ms = NULL",
-            &[&principal_id, &token_sha256, &role, &created_ms],
-        )?;
-        Ok(())
-    }
-
-    fn revoke_api_principal(
-        &mut self,
-        principal_id: &str,
-        now_ms: u64,
-    ) -> Result<bool, PgStoreError> {
-        let rev_ms = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        let n = self.lock().execute(
-            "UPDATE api_principals SET revoked_at_ms = $2 \
-             WHERE principal_id = $1 AND revoked_at_ms IS NULL",
-            &[&principal_id, &rev_ms],
-        )?;
-        Ok(n > 0)
-    }
-
-    fn load_api_principal_by_token_hash(
-        &self,
-        token_sha256: &str,
-    ) -> Result<Option<ApiPrincipalRecord>, PgStoreError> {
-        // Lookup by hash only; the record never carries the token back.
-        let row = self.lock().query_opt(
-            "SELECT principal_id, role, created_at_ms, revoked_at_ms \
-             FROM api_principals WHERE token_sha256 = $1",
-            &[&token_sha256],
-        )?;
-        Ok(row.as_ref().map(Self::row_to_api_principal))
-    }
-
-    fn load_api_principals(&self) -> Result<Vec<ApiPrincipalRecord>, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT principal_id, role, created_at_ms, revoked_at_ms \
-             FROM api_principals ORDER BY principal_id",
-            &[],
-        )?;
-        Ok(rows.iter().map(Self::row_to_api_principal).collect())
-    }
-}
-
-impl PgVerifierStore {
-    fn row_to_cert_principal(row: &postgres::Row) -> CertPrincipalRecord {
-        CertPrincipalRecord {
-            principal_id: row.get(0),
-            role: row.get(1),
-            created_at_ms: row.get::<_, i64>(2) as u64,
-            revoked_at_ms: row.get::<_, Option<i64>>(3).map(|v| v as u64),
-            // FAIL-CLOSED read (matches the SQLite backend, Copilot #857): a NEGATIVE
-            // stored `not_after_ms` — only reachable via corruption, since the write
-            // path refuses `> i64::MAX` — maps to `Some(0)` ("expired at epoch"), so a
-            // tampered expiry can only make a cert MORE restricted, never a huge
-            // never-expiring value. `u64::try_from` fails only for negatives → 0.
-            not_after_ms: row
-                .get::<_, Option<i64>>(4)
-                .map(|v| u64::try_from(v).unwrap_or(0)),
-        }
-    }
-}
-
-impl CertPrincipalStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn register_cert_principal(
-        &mut self,
-        principal_id: &str,
-        cert_sha256: &str,
-        role: &str,
-        not_after_ms: Option<u64>,
-        now_ms: u64,
-    ) -> Result<(), PgStoreError> {
-        let created_ms = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // Refuse a `not_after_ms > i64::MAX` (never truncate to a negative that would
-        // read back as a huge never-expiring value — a fail-OPEN expiry). Bounded in
-        // practice (~292M years past epoch).
-        let not_after_i64 = match not_after_ms {
-            Some(v) => Some(i64::try_from(v).map_err(|_| PgStoreError::OutOfDomain {
-                field: "not_after_ms",
-                value: v,
-            })?),
-            None => None,
-        };
-        // A `cert_sha256` already pinned to a DIFFERENT principal violates the UNIQUE
-        // constraint (surfaces as a driver error — one cert, one principal).
-        self.lock().execute(
-            "INSERT INTO cert_principals \
-                 (principal_id, cert_sha256, role, created_at_ms, revoked_at_ms, not_after_ms) \
-             VALUES ($1, $2, $3, $4, NULL, $5) \
-             ON CONFLICT (principal_id) DO UPDATE SET \
-                 cert_sha256   = EXCLUDED.cert_sha256, \
-                 role          = EXCLUDED.role, \
-                 created_at_ms = EXCLUDED.created_at_ms, \
-                 revoked_at_ms = NULL, \
-                 not_after_ms  = EXCLUDED.not_after_ms",
-            &[
-                &principal_id,
-                &cert_sha256,
-                &role,
-                &created_ms,
-                &not_after_i64,
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn revoke_cert_principal(
-        &mut self,
-        principal_id: &str,
-        now_ms: u64,
-    ) -> Result<bool, PgStoreError> {
-        let rev_ms = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        let n = self.lock().execute(
-            "UPDATE cert_principals SET revoked_at_ms = $2 \
-             WHERE principal_id = $1 AND revoked_at_ms IS NULL",
-            &[&principal_id, &rev_ms],
-        )?;
-        Ok(n > 0)
-    }
-
-    fn load_cert_principal_by_fingerprint(
-        &self,
-        cert_sha256: &str,
-    ) -> Result<Option<CertPrincipalRecord>, PgStoreError> {
-        let row = self.lock().query_opt(
-            "SELECT principal_id, role, created_at_ms, revoked_at_ms, not_after_ms \
-             FROM cert_principals WHERE cert_sha256 = $1",
-            &[&cert_sha256],
-        )?;
-        Ok(row.as_ref().map(Self::row_to_cert_principal))
-    }
-
-    fn load_cert_principals(&self) -> Result<Vec<CertPrincipalRecord>, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT principal_id, role, created_at_ms, revoked_at_ms, not_after_ms \
-             FROM cert_principals ORDER BY principal_id",
-            &[],
-        )?;
-        Ok(rows.iter().map(Self::row_to_cert_principal).collect())
-    }
-}
-
-impl PgVerifierStore {
-    fn row_to_fabric_asset(row: &postgres::Row) -> FabricAsset {
-        // Same lenient decode as the SQLite backend: an undecodable enum falls back
-        // (asset_type → Unknown, profile → Custom), a bad metadata blob → empty map —
-        // never a panic or a skipped row.
-        let asset_type_s: String = row.get(1);
-        let profile_s: String = row.get(3);
-        let meta_s: String = row.get(6);
-        FabricAsset {
-            asset_id: row.get(0),
-            asset_type: serde_json::from_str(&asset_type_s).unwrap_or(AssetType::Unknown),
-            display_name: row.get(2),
-            kinematic_profile: serde_json::from_str(&profile_s)
-                .unwrap_or(KinematicProfileType::Custom),
-            // Fail-CLOSED read (matches the save-path guard + `not_after_ms`): a
-            // corrupt NEGATIVE stored timestamp maps to 0 rather than wrapping to a
-            // huge u64 that would corrupt ordering/visibility. `u64::try_from` fails
-            // only for negatives → 0.
-            registered_at_ms: u64::try_from(row.get::<_, i64>(4)).unwrap_or(0),
-            last_seen_ms: u64::try_from(row.get::<_, i64>(5)).unwrap_or(0),
-            metadata: serde_json::from_str(&meta_s).unwrap_or_default(),
-        }
-    }
-}
-
-impl FabricAssetStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn save_fabric_asset(&self, asset: &FabricAsset) -> Result<(), PgStoreError> {
-        // Fail-closed on out-of-domain timestamps (the #936 lesson) rather than
-        // wrapping a u64 to a negative BIGINT.
-        let reg_ms =
-            i64::try_from(asset.registered_at_ms).map_err(|_| PgStoreError::OutOfDomain {
-                field: "registered_at_ms",
-                value: asset.registered_at_ms,
-            })?;
-        let last_ms = i64::try_from(asset.last_seen_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "last_seen_ms",
-            value: asset.last_seen_ms,
-        })?;
-        // The enum fields + metadata map are JSON-serialized into TEXT, exactly as
-        // the SQLite backend (same encoding → same bytes round-trip).
-        let asset_type = serde_json::to_string(&asset.asset_type).unwrap_or_default();
-        let profile = serde_json::to_string(&asset.kinematic_profile).unwrap_or_default();
-        let metadata = serde_json::to_string(&asset.metadata).unwrap_or_else(|_| "{}".to_string());
-        // Upsert by asset_id (SQLite: INSERT OR REPLACE).
-        self.lock().execute(
-            "INSERT INTO fabric_assets \
-                 (asset_id, asset_type, display_name, kinematic_profile, registered_at_ms, \
-                  last_seen_ms, metadata_json) \
-             VALUES ($1, $2, $3, $4, $5, $6, $7) \
-             ON CONFLICT (asset_id) DO UPDATE SET \
-                 asset_type        = EXCLUDED.asset_type, \
-                 display_name      = EXCLUDED.display_name, \
-                 kinematic_profile = EXCLUDED.kinematic_profile, \
-                 registered_at_ms  = EXCLUDED.registered_at_ms, \
-                 last_seen_ms      = EXCLUDED.last_seen_ms, \
-                 metadata_json     = EXCLUDED.metadata_json",
-            &[
-                &asset.asset_id,
-                &asset_type,
-                &asset.display_name,
-                &profile,
-                &reg_ms,
-                &last_ms,
-                &metadata,
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn load_fabric_assets(&self) -> Result<Vec<FabricAsset>, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT asset_id, asset_type, display_name, kinematic_profile, registered_at_ms, \
-                    last_seen_ms, metadata_json \
-             FROM fabric_assets ORDER BY registered_at_ms, asset_id",
-            &[],
-        )?;
-        Ok(rows.iter().map(Self::row_to_fabric_asset).collect())
-    }
-}
-
-/// The ordered column list for an `ota_campaigns` SELECT, shared by the three read
-/// paths so [`PgVerifierStore::row_to_campaign`]'s positional `row.get(..)` indices
-/// stay in lock-step with the projection.
-const CAMPAIGN_COLUMNS: &str = "campaign_id, artifact_digest, artifact_version, cohorts_json, \
-                                stages_json, stage_index, rollout_percent, state, halt_reason, \
-                                created_at_ms, updated_at_ms, artifact_signature_b64, \
-                                uptane_metadata_json";
-
-impl PgVerifierStore {
-    /// Decode one `ota_campaigns` row into a [`Campaign`], FAIL-CLOSED exactly as
-    /// the SQLite backend's `map_campaign_row`: a malformed `cohorts`/`stages` blob,
-    /// an unknown `state`/`halt_reason` token, or a `stage_index` out of range for
-    /// the schedule is an error, never a silently-defaulted `Campaign` the engine
-    /// could later index out of bounds. Timestamps read fail-closed (a negative
-    /// stored BIGINT → 0, never a wrap to a huge `u64`), matching the FabricAsset
-    /// read path and the #936/#938 lesson.
-    fn row_to_campaign(row: &postgres::Row) -> Result<Campaign, PgStoreError> {
-        let cohorts_json: String = row.get(3);
-        let stages_json: String = row.get(4);
-        let state_s: String = row.get(7);
-        let halt_s: Option<String> = row.get(8);
-
-        // A malformed stored blob is stored-row CORRUPTION (the read-path inverse of
-        // an encode failure), classified like SQLite's `json_decode_err` — NOT
-        // `Encode`, which is strictly the write-path direction.
-        let cohorts: Vec<String> = serde_json::from_str(&cohorts_json)
-            .map_err(|e| PgStoreError::CorruptCampaignRow(format!("cohorts_json decode: {e}")))?;
-        let stages: Vec<u8> = serde_json::from_str(&stages_json)
-            .map_err(|e| PgStoreError::CorruptCampaignRow(format!("stages_json decode: {e}")))?;
-        let state = CampaignState::parse(&state_s)
-            .ok_or_else(|| PgStoreError::CorruptCampaignRow(format!("state token {state_s:?}")))?;
-        let halt_reason = match halt_s {
-            Some(s) => Some(HaltReason::parse(&s).ok_or_else(|| {
-                PgStoreError::CorruptCampaignRow(format!("halt_reason token {s:?}"))
-            })?),
-            None => None,
-        };
-
-        // CHECKED numeric conversions — a tampered row must fail closed, never wrap a
-        // negative/huge BIGINT into a bogus `usize`/`u8` the engine could index out
-        // of bounds.
-        let stage_index_raw: i64 = row.get(5);
-        let stage_index = usize::try_from(stage_index_raw).map_err(|_| {
-            PgStoreError::CorruptCampaignRow(format!("stage_index {stage_index_raw}"))
-        })?;
-        // The current stage must index into the schedule (true for every REACHABLE
-        // campaign) — out of range is corruption.
-        if stage_index >= stages.len() {
-            return Err(PgStoreError::CorruptCampaignRow(format!(
-                "stage_index {stage_index_raw} out of range for {} stage(s)",
-                stages.len()
-            )));
-        }
-        let rollout_raw: i64 = row.get(6);
-        let rollout_percent = u8::try_from(rollout_raw)
-            .ok()
-            .filter(|p| *p <= 100)
-            .ok_or_else(|| {
-                PgStoreError::CorruptCampaignRow(format!("rollout_percent {rollout_raw}"))
-            })?;
-
-        Ok(Campaign {
-            campaign_id: row.get(0),
-            artifact_digest: row.get(1),
-            artifact_version: row.get(2),
-            cohorts,
-            stages,
-            stage_index,
-            rollout_percent,
-            state,
-            halt_reason,
-            // Fail-closed timestamp reads (negative → 0), matching the save-path
-            // OutOfDomain guard and the FabricAsset read path.
-            created_at_ms: u64::try_from(row.get::<_, i64>(9)).unwrap_or(0),
-            updated_at_ms: u64::try_from(row.get::<_, i64>(10)).unwrap_or(0),
-            artifact_signature_b64: row.get(11),
-            uptane_metadata_json: row.get(12),
-        })
-    }
-
-    fn row_to_node_artifact_status(row: &postgres::Row) -> NodeArtifactStatus {
-        NodeArtifactStatus {
-            node_id: row.get(0),
-            applied_digest: row.get(1),
-            campaign_id: row.get(2),
-            artifact_version: row.get(3),
-            // Fail-closed read (negative → 0), as everywhere else in this backend.
-            reported_at_ms: u64::try_from(row.get::<_, i64>(4)).unwrap_or(0),
-            attested: row.get(5),
-        }
-    }
-}
-
-impl OtaCampaignStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn insert_campaign(&mut self, campaign: &Campaign) -> Result<(), PgStoreError> {
-        // Fail-closed on encode (a bad cohorts/stages blob is load-bearing — stages
-        // bounds stage_index) and on out-of-domain timestamps (the #936 lesson).
-        let cohorts_json =
-            serde_json::to_string(&campaign.cohorts).map_err(PgStoreError::Encode)?;
-        let stages_json = serde_json::to_string(&campaign.stages).map_err(PgStoreError::Encode)?;
-        let stage_index =
-            i64::try_from(campaign.stage_index).map_err(|_| PgStoreError::OutOfDomain {
-                field: "stage_index",
-                value: campaign.stage_index as u64,
-            })?;
-        let created_ms =
-            i64::try_from(campaign.created_at_ms).map_err(|_| PgStoreError::OutOfDomain {
-                field: "created_at_ms",
-                value: campaign.created_at_ms,
-            })?;
-        let updated_ms =
-            i64::try_from(campaign.updated_at_ms).map_err(|_| PgStoreError::OutOfDomain {
-                field: "updated_at_ms",
-                value: campaign.updated_at_ms,
-            })?;
-        let rollout = i64::from(campaign.rollout_percent);
-        let state = campaign.state.as_str();
-        let halt_reason = campaign.halt_reason.map(|r| r.as_str());
-        // PLAIN INSERT — a duplicate `campaign_id` violates the PRIMARY KEY and
-        // surfaces as a driver error (the storage contract's "duplicate id
-        // conflicts, never a silent overwrite"), exactly as the SQLite backend's
-        // plain INSERT. The audit-chaining SQLite additionally performs on create is
-        // an inherent concern OUTSIDE this storage contract. Autocommit: the failed
-        // INSERT is its own implicit tx, leaving the connection usable.
-        self.lock().execute(
-            "INSERT INTO ota_campaigns \
-                 (campaign_id, artifact_digest, artifact_version, cohorts_json, stages_json, \
-                  stage_index, rollout_percent, state, halt_reason, created_at_ms, updated_at_ms, \
-                  artifact_signature_b64, uptane_metadata_json) \
-             VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13)",
-            &[
-                &campaign.campaign_id,
-                &campaign.artifact_digest,
-                &campaign.artifact_version,
-                &cohorts_json,
-                &stages_json,
-                &stage_index,
-                &rollout,
-                &state,
-                &halt_reason,
-                &created_ms,
-                &updated_ms,
-                &campaign.artifact_signature_b64,
-                &campaign.uptane_metadata_json,
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn load_campaign(&self, campaign_id: &str) -> Result<Option<Campaign>, PgStoreError> {
-        let sql = format!("SELECT {CAMPAIGN_COLUMNS} FROM ota_campaigns WHERE campaign_id = $1");
-        let rows = self.lock().query(&sql, &[&campaign_id])?;
-        match rows.first() {
-            Some(row) => Ok(Some(Self::row_to_campaign(row)?)),
-            None => Ok(None),
-        }
-    }
-
-    fn load_campaigns(&self) -> Result<Vec<Campaign>, PgStoreError> {
-        // Newest first: created_at_ms DESC, then campaign_id ASC (deterministic ties).
-        let sql = format!(
-            "SELECT {CAMPAIGN_COLUMNS} FROM ota_campaigns \
-             ORDER BY created_at_ms DESC, campaign_id ASC"
-        );
-        let rows = self.lock().query(&sql, &[])?;
-        rows.iter().map(Self::row_to_campaign).collect()
-    }
-
-    fn load_active_campaigns(&self) -> Result<Vec<Campaign>, PgStoreError> {
-        // Only Staged/Rolling (the sweep-monitor set), oldest first: created_at_ms
-        // ASC, then campaign_id ASC. The state tokens match `CampaignState::as_str`.
-        let sql = format!(
-            "SELECT {CAMPAIGN_COLUMNS} FROM ota_campaigns \
-             WHERE state IN ('staged', 'rolling') \
-             ORDER BY created_at_ms ASC, campaign_id ASC"
-        );
-        let rows = self.lock().query(&sql, &[])?;
-        rows.iter().map(Self::row_to_campaign).collect()
-    }
-
-    fn upsert_node_artifact_status(&mut self, st: &NodeArtifactStatus) -> Result<(), PgStoreError> {
-        let reported_ms =
-            i64::try_from(st.reported_at_ms).map_err(|_| PgStoreError::OutOfDomain {
-                field: "reported_at_ms",
-                value: st.reported_at_ms,
-            })?;
-        // MONOTONIC upsert with attested-per-digest carry, byte-for-byte the SQLite
-        // backend's semantics: the `WHERE excluded.reported_at_ms >= …` makes a
-        // stale report a no-op; `attested` can only be SET or preserved-for-the-
-        // same-digest, never cleared by a later unsigned same-digest report — but a
-        // DIFFERENT digest resets it (the RHS reads the pre-update row).
-        self.lock().execute(
-            "INSERT INTO node_artifact_status \
-                 (node_id, applied_digest, campaign_id, artifact_version, reported_at_ms, attested) \
-             VALUES ($1, $2, $3, $4, $5, $6) \
-             ON CONFLICT (node_id) DO UPDATE SET \
-                 applied_digest   = EXCLUDED.applied_digest, \
-                 campaign_id      = EXCLUDED.campaign_id, \
-                 artifact_version = EXCLUDED.artifact_version, \
-                 reported_at_ms   = EXCLUDED.reported_at_ms, \
-                 attested         = EXCLUDED.attested \
-                    OR (node_artifact_status.attested \
-                        AND node_artifact_status.applied_digest = EXCLUDED.applied_digest) \
-               WHERE EXCLUDED.reported_at_ms >= node_artifact_status.reported_at_ms",
-            &[
-                &st.node_id,
-                &st.applied_digest,
-                &st.campaign_id,
-                &st.artifact_version,
-                &reported_ms,
-                &st.attested,
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn load_node_artifact_statuses(&self) -> Result<Vec<NodeArtifactStatus>, PgStoreError> {
-        // Newest report first: reported_at_ms DESC, then node_id ASC.
-        let rows = self.lock().query(
-            "SELECT node_id, applied_digest, campaign_id, artifact_version, reported_at_ms, attested \
-             FROM node_artifact_status ORDER BY reported_at_ms DESC, node_id ASC",
-            &[],
-        )?;
-        Ok(rows.iter().map(Self::row_to_node_artifact_status).collect())
-    }
-}
-
-impl PgVerifierStore {
-    fn row_to_av_subsystem(row: &postgres::Row) -> AvSubsystemRecord {
-        AvSubsystemRecord {
-            node_id: row.get(0),
-            subsystem_type: row.get(1),
-            hardware_id: row.get(2),
-            confidence_floor: row.get(3),
-            // Fail-closed reads (negative → 0), as everywhere else in this backend.
-            last_telemetry_ms: u64::try_from(row.get::<_, i64>(4)).unwrap_or(0),
-            recovery_streak_count: u32::try_from(row.get::<_, i64>(5)).unwrap_or(0),
-            recovery_streak_start_ms: u64::try_from(row.get::<_, i64>(6)).unwrap_or(0),
-        }
-    }
-}
-
-impl AvSubsystemStore for PgVerifierStore {
-    type Error = PgStoreError;
-
-    fn register_av_subsystem_meta(
-        &self,
-        node_id: &str,
-        subsystem_type: &str,
-        hardware_id: &str,
-        confidence_floor: f64,
-        initial_telemetry_ms: u64,
-    ) -> Result<(), PgStoreError> {
-        let tel_ms =
-            i64::try_from(initial_telemetry_ms).map_err(|_| PgStoreError::OutOfDomain {
-                field: "initial_telemetry_ms",
-                value: initial_telemetry_ms,
-            })?;
-        // Faithful INSERT-OR-REPLACE: a re-register overwrites the row AND resets the
-        // streak counters to 0 (SQLite's INSERT OR REPLACE drops the old row, so the
-        // streak columns fall back to DEFAULT 0). The `DO UPDATE` therefore explicitly
-        // zeroes the streak — they are not in the insert projection.
-        self.lock().execute(
-            "INSERT INTO av_subsystem_meta \
-                 (node_id, subsystem_type, hardware_id, confidence_floor, last_telemetry_ms, \
-                  recovery_streak_count, recovery_streak_start_ms) \
-             VALUES ($1, $2, $3, $4, $5, 0, 0) \
-             ON CONFLICT (node_id) DO UPDATE SET \
-                 subsystem_type           = EXCLUDED.subsystem_type, \
-                 hardware_id              = EXCLUDED.hardware_id, \
-                 confidence_floor         = EXCLUDED.confidence_floor, \
-                 last_telemetry_ms        = EXCLUDED.last_telemetry_ms, \
-                 recovery_streak_count    = 0, \
-                 recovery_streak_start_ms = 0",
-            &[
-                &node_id,
-                &subsystem_type,
-                &hardware_id,
-                &confidence_floor,
-                &tel_ms,
-            ],
-        )?;
-        Ok(())
-    }
-
-    fn load_av_confidence_floor(&self, node_id: &str) -> Result<Option<f64>, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT confidence_floor FROM av_subsystem_meta WHERE node_id = $1",
-            &[&node_id],
-        )?;
-        Ok(rows.first().map(|r| r.get::<_, f64>(0)))
-    }
-
-    fn touch_av_telemetry_timestamp(&self, node_id: &str, now_ms: u64) -> Result<(), PgStoreError> {
-        let now = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // No-op if the node is unregistered (0 rows updated), exactly as SQLite.
-        self.lock().execute(
-            "UPDATE av_subsystem_meta SET last_telemetry_ms = $1 WHERE node_id = $2",
-            &[&now, &node_id],
-        )?;
-        Ok(())
-    }
-
-    fn get_last_telemetry_timestamp(&self, node_id: &str) -> Result<u64, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT last_telemetry_ms FROM av_subsystem_meta WHERE node_id = $1",
-            &[&node_id],
-        )?;
-        // Absent → 0 (matches SQLite's QueryReturnedNoRows → 0); negative → 0.
-        Ok(rows
-            .first()
-            .map(|r| u64::try_from(r.get::<_, i64>(0)).unwrap_or(0))
-            .unwrap_or(0))
-    }
-
-    fn load_all_registered_av_node_ids(&self) -> Result<Vec<String>, PgStoreError> {
-        // Ordered by node_id (deterministic; the contract sorts anyway).
-        let rows = self.lock().query(
-            "SELECT node_id FROM av_subsystem_meta ORDER BY node_id",
-            &[],
-        )?;
-        Ok(rows.iter().map(|r| r.get::<_, String>(0)).collect())
-    }
-
-    fn load_av_subsystems(&self) -> Result<Vec<AvSubsystemRecord>, PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT node_id, subsystem_type, hardware_id, confidence_floor, last_telemetry_ms, \
-                    recovery_streak_count, recovery_streak_start_ms \
-             FROM av_subsystem_meta ORDER BY node_id",
-            &[],
-        )?;
-        Ok(rows.iter().map(Self::row_to_av_subsystem).collect())
-    }
-
-    fn load_recovery_streak(&self, node_id: &str) -> Result<(u32, u64), PgStoreError> {
-        let rows = self.lock().query(
-            "SELECT recovery_streak_count, recovery_streak_start_ms \
-             FROM av_subsystem_meta WHERE node_id = $1",
-            &[&node_id],
-        )?;
-        // Absent → (0, 0); fail-closed reads (negative → 0).
-        Ok(rows
-            .first()
-            .map(|r| {
-                (
-                    u32::try_from(r.get::<_, i64>(0)).unwrap_or(0),
-                    u64::try_from(r.get::<_, i64>(1)).unwrap_or(0),
-                )
-            })
-            .unwrap_or((0, 0)))
-    }
-
-    fn reset_recovery_streak(&self, node_id: &str, now_ms: u64) -> Result<(), PgStoreError> {
-        let now = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        self.lock().execute(
-            "UPDATE av_subsystem_meta \
-                SET recovery_streak_count = 0, recovery_streak_start_ms = 0, \
-                    last_telemetry_ms = $1 \
-              WHERE node_id = $2",
-            &[&now, &node_id],
-        )?;
-        Ok(())
-    }
-
-    fn reset_recovery_streak_preserving_telemetry(
-        &self,
-        node_id: &str,
-    ) -> Result<(), PgStoreError> {
-        self.lock().execute(
-            "UPDATE av_subsystem_meta \
-                SET recovery_streak_count = 0, recovery_streak_start_ms = 0 \
-              WHERE node_id = $1",
-            &[&node_id],
-        )?;
-        Ok(())
-    }
-
-    fn increment_recovery_streak(&self, node_id: &str, now_ms: u64) -> Result<u32, PgStoreError> {
-        let now = i64::try_from(now_ms).map_err(|_| PgStoreError::OutOfDomain {
-            field: "now_ms",
-            value: now_ms,
-        })?;
-        // Single-statement UPDATE … RETURNING: `start_ms` stamps only on the 0→1 edge
-        // (the CASE mirrors SQLite exactly), telemetry advances, and the new count is
-        // returned. Zero rows returned = the node has no row → fail-closed
-        // `AvNodeNotRegistered` (SQLite surfaces this as QueryReturnedNoRows, the
-        // in-memory backend as NodeNotRegistered).
-        let rows = self.lock().query(
-            "UPDATE av_subsystem_meta \
-                SET recovery_streak_count = recovery_streak_count + 1, \
-                    recovery_streak_start_ms = CASE \
-                        WHEN recovery_streak_count = 0 THEN $1 \
-                        ELSE recovery_streak_start_ms END, \
-                    last_telemetry_ms = $1 \
-              WHERE node_id = $2 \
-              RETURNING recovery_streak_count",
-            &[&now, &node_id],
-        )?;
-        match rows.first() {
-            Some(r) => Ok(u32::try_from(r.get::<_, i64>(0)).unwrap_or(0)),
-            None => Err(PgStoreError::AvNodeNotRegistered(node_id.to_string())),
-        }
-    }
-}
+mod av_subsystem;
+mod cert_principals;
+mod epoch;
+mod fabric;
+mod federation;
+mod nodes;
+mod operators;
+mod ota_campaigns;
+mod posture;
+mod principals;
